+++ src/main/java/flab/rocket_market/dto/UpdateProductRequest.java
@@ -1,6 +1,5 @@
-package flab.rocket_market.controller.dto;
+package flab.rocket_market.dto;
 
-import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -13,16 +12,12 @@ import java.math.BigDecimal;
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor
-public class RegisterProductRequest {
+public class UpdateProductRequest {
 
-    @NotBlank(message = "{product.name.notblank}")
+    @NotNull(message = "{product.id.notnull}")
+    private Long productId;
     private String name;
-
     private String description;
-
-    @NotNull(message = "{product.price.notnull}")
     private BigDecimal price;
-
-    @NotNull(message = "{product.category.notnull}")
     private Long categoryId;
 }
